added delete archival record via it's pid
https://zimlab.uni-graz.at/gams5/gams-management/gams-cms/gams-api/-/work_items/85

+++ src/main/java/org/ddh/gamsapi/domain/ArchivalRecord/ArchivalRecordService.java
@@ -55,6 +55,11 @@ public class ArchivalRecordService implements IArchivalRecordService {
 
   @Override
   public void deleteById(String archivalRecordPid) {
+    if(!archivalRecordRepository.existsById(archivalRecordPid)){
+      throw new ArchivalRecordNotFoundException(
+          "Cannot delete archival record with pid: " + archivalRecordPid + " The archival record does not exist."
+      );
+    }
     archivalRecordRepository.deleteById(archivalRecordPid);
   }
